+++ todo_app/tests/test_e2e.py
@@ -1,0 +1,3 @@
+def test_task_journey(driver, app_with_temp_board):
+    driver.get('http://localhost:5000/')
+    assert driver.title == 'To-Do App'
